fixes: #491 move wallet error functions to errorutils and rename to be clearer on purpose
