+++ bin/settlement-submit/main.go
@@ -12,8 +12,8 @@ import (
 	"strings"
 
 	"github.com/brave-intl/bat-go/settlement"
+	errorutils "github.com/brave-intl/bat-go/utils/errors"
 	"github.com/brave-intl/bat-go/utils/formatters"
-	"github.com/brave-intl/bat-go/utils/wallet"
 	"github.com/brave-intl/bat-go/utils/wallet/provider/uphold"
 	log "github.com/sirupsen/logrus"
 )
@@ -89,7 +89,7 @@ func main() {
 
 		err = settlement.SubmitPreparedTransaction(settlementWallet, settlementTransaction)
 		if err != nil {
-			if wallet.IsInvalidDestination(err) {
+			if errorutils.IsErrInvalidDestination(err) {
 				log.Println(err)
 				continue
 			}
