+++ utils/wallet/wallet.go
@@ -85,57 +85,3 @@ type Wallet interface {
 	// ListTransactions for this wallet, limit number of transactions returned
 	ListTransactions(limit int, startDate time.Time) ([]TransactionInfo, error)
 }
-
-// IsNotFound is a helper method for determining if an error indicates a missing resource
-func IsNotFound(err error) bool {
-	type notFound interface {
-		NotFoundError() bool
-	}
-	te, ok := err.(notFound)
-	return ok && te.NotFoundError()
-}
-
-// IsInvalidDestination is a helper method for determining if an error indicates an invalid destination
-func IsInvalidDestination(err error) bool {
-	type invalidDestination interface {
-		InvalidDestination() bool
-	}
-	te, ok := err.(invalidDestination)
-	return ok && te.InvalidDestination()
-}
-
-// IsInsufficientBalance is a helper method for determining if an error indicates insufficient balance
-func IsInsufficientBalance(err error) bool {
-	type insufficientBalance interface {
-		InsufficientBalance() bool
-	}
-	te, ok := err.(insufficientBalance)
-	return ok && te.InsufficientBalance()
-}
-
-// IsUnauthorized is a helper method for determining if an error indicates the wallet unauthorized
-func IsUnauthorized(err error) bool {
-	type unauthorized interface {
-		Unauthorized() bool
-	}
-	te, ok := err.(unauthorized)
-	return ok && te.Unauthorized()
-}
-
-// IsInvalidSignature is a helper method for determining if an error indicates there was an invalid signature
-func IsInvalidSignature(err error) bool {
-	type invalidSignature interface {
-		InvalidSignature() bool
-	}
-	te, ok := err.(invalidSignature)
-	return ok && te.InvalidSignature()
-}
-
-// AlreadyExists is a helper method for determining if an error indicates the resource already exists
-func AlreadyExists(err error) bool {
-	type alreadyExists interface {
-		AlreadyExistsError() bool
-	}
-	te, ok := err.(alreadyExists)
-	return ok && te.AlreadyExistsError()
-}
